requirements.txt
discord.py-self
iniconfig
pet-pet-gif
requests
